Refactor card decorator and related specs
